Move EntityRenamer
